feat: wip, migration of generic form
